almost finished with the instructions, last 10 remain

+++ EMoo/mem/ram.h
@@ -21,15 +21,15 @@ public:
     void loadFromFile(const char *filename, uint32_t address);
     void write(uint32_t address, uint8_t* source, uint32_t length);
 
-    uint8_t read8(uint8_t address);
-    uint16_t read16(uint8_t address);
+    uint8_t read8(uint32_t address);
+    uint16_t read16(uint32_t address);
 };
 
-inline uint8_t Ram::read8(uint8_t address) {
+inline uint8_t Ram::read8(uint32_t address) {
     return buffer[address];
 }
 
-inline uint16_t Ram::read16(uint8_t address) {
+inline uint16_t Ram::read16(uint32_t address) {
     return *((uint16_t*)(buffer + address));
 }
